requirements.txt
flask>=3.0.0
dash>=2.17.0
dash-bootstrap-components>=1.6.0
plotly>=5.22.0
pandas>=2.2.0
numpy>=1.26.0
pymongo>=4.8.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
reportlab>=4.2.0
openpyxl>=3.1.0
xlsxwriter>=3.2.0
pdfplumber>=0.11.0
requests>=2.32.0